requirements.txt
tensorflow==2.13.0
streamlit==1.25.0
pillow==9.5.0
numpy==1.25.2